pyproject.toml
[tool.poetry]
name = "winds-mobi-api"
description = "Python standalone asyncIO API that provides different endpoints to get weather data from winds.mobi mongodb."
version = "0.0.0"
authors = ["winds.mobi"]
license = "	AGPL-3.0-only"
packages = [
    { include = "winds_mobi_api" },
]

[tool.poetry.dependencies]
python = "3.11.9"

aiocache = "0.12.2"
fastapi = "0.111.0"
motor = "3.5.1"
opentelemetry-distro = {extras = ["otlp"], version = "0.46b0"}
opentelemetry-instrumentation = "0.46b0"
opentelemetry-instrumentation-fastapi = "0.46b0"
opentelemetry-instrumentation-pymongo = "0.46b0"
orjson = "3.10.6"
parse-accept-language = "0.1.2"
pyaml = "24.4.0"
pydantic = "2.8.2"
pydantic-settings = "2.3.4"
scipy = "1.14.0"
sentry-asgi = "0.2.0"
sentry-sdk = "2.9.0"
stop-words = "2018.7.23"
ujson = "5.10.0" # Optional dependency used by aiocache

[tool.poetry.group.dev.dependencies]
black = "24.4.2"
flake8 = "7.1.0"
isort = "5.13.2"
locust = "*"
pytest = "8.2.2"
python-dotenv = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
